requirements.txt
Django==4.2.16
djangorestframework==3.15.2
psycopg[binary]
flake8==7.1.1